+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for private CAPZ cluster
+
 ## [1.20.0] - 2024-08-19
 
 ### Changed